requirements.txt
discord.py
PyNaCl
discord.py[voice]
pip
youtube_dl
colorlog